+++ app/Enums/RuleType.php
@@ -4,8 +4,7 @@ namespace App\Enums;
 
 enum RuleType: string
 {
-    case TeamSize = 'team_size';
-    case GroupSize = 'group_size';
+    case Size = 'size';
     case BannedPair = 'banned_pair';
     case RepeatPair = 'repeat_pair';
     case PreferredPair = 'preferred_pair';